Add direct support for "hash-storage" mode
### The WHY:

This change is part of a series with the overall objective of
further improving the already very capable GLM in order to
offer a smoother migration path for more existing GoL usage
patterns, mainly by ensuring a tighter compatibility with
<Getopt::Long>, so as to minimize changes to legacy code
that may be needed during any initial migration.

In the end, it will hopefully become a true no-brainer for any
existing GoL consumer to add exactly 6 characters, i.e. "::More"
on exactly one one line of code on top of any historical script
for them to immediately start enjoying "More" goodness, and that
without having to give to give up any "Long" existing goodness,
in as much as possible, and especially in terms of DRY & DWIM.

At least that's the holly grail that GLM has set itself up from
the beginning; and it's probably not far from there at this point.

Anyhow, there still appear to remain some concrete scenarios
where that doesn't quite happen, yet. Issues #1 and #3 described
two such cases.

Previous PR [#4] was focused on resolving issue #3. Once that was
out way, it became trivial to fix issue #1, which this commit aims
to achieve.

From what I can see, the only other hurdle (to some of those easy-peasy
migration scenarios) that remain at this time is a design choice
described under issue #5, which needs to be discussed before its
corresponding simple fix can be submitted, if at all.

### The WHAT:

#### PREVIOUS BEHAVIOUR

As a reminder, when called with the "hash-storage" mode signature,
the previous behavior of GLM's `GetoptsFromArray` wrapper was to
abstain from doing anything itself as soon as it detected this
situation, effectively reducing itself to nothing more than a
glorified stack frame filler :-)

That approach works, as long as there are no "OptSpec" objects
around (i.e. the very initial point in a GoL => GLM migration).

Once you try to add an OptSpec, though (which may be months
after the initial migration of a script), GoL (which naturally
has no knowledge of OptSpec), would either immediately throw an
error, or at least get utterly confused because it would try
treating that OptSpec as an ordinary -HASH ref- "linkage"
for the associated optspec (HASH ref linkage is quite a
valid one for some cases).

Anyhow, you get the point...

#### NEW BEHAVIOUR

The new behavior is actually a uniformisation / harmonisation,
both functionally and code-wise.

Basically, we no longer refrain ourselves from venturing into
doing the usual GLM things (OptSpec extraction, shell completion,
automatic help message generation, ...) in "hash-storage" mode,
which is no longer treated in any special manner; well almost...

In other words, no short-circuiting takes place any longer, meaning
every single usage pattern goes through the same code path, greatly
simplifying the overall story.

The only "special treatment" received by the "hash-storage"
mode has to do with the call frame setup/plumbing (detecting the mode,
stowing away the hash-ref, and making it available to our wrapped
counterpart when its time to call it.

### The HOW:

I had already added a few test cases to cover this fix (for gh #1)
as well as the previous change that handles gh #3 and unblocks this
fix.

Some of those tests  already passing, some others had to be marked
as "TODO" to prevent any adverse impact on current builds on Master.

The actual changes to the code related to this fix itself were limited
to just 3-4 lines, and were focused on the step where we detect the
"hash-storage" mode call signature.

As mentioned earlier, we no longer short-circuit and bypass our own
code path upon detecting this mode. Instead, we just save the hash
reference detected to be that "storage" so that it can be re-inserted
at the same point to the list of arguments when it is time to call
our wrapped counterpart.

Since a suitable mechanism already existed in the code (@go_opts_spec),
all we had to do was to pull up its declaration to happen earlier on
so that we could make use of it.

That's pretty much it. So far, it's quite straight forward, EXCEPT:

One question may perhaps remain, discussed in detail on the sidelines
of #1, where these same changes were given as a patch.

It would be too long to detail in its entirety here, but it's about
the check that actually detects the hash-storage mode call signature.

The previous code checked : `( $_[0] eq 'HASH')`

Hence it was limited to detecting pure-hashes only; as opposed to GoL
itself which tries to allow any hashy ref it receives, including objects,
as a "hash-storage" destination, as long as the underlying object storage
remained to be a form of HASH.

The difference is subtle, but may be important at times. Therefore,
I sought to more closely mimic GoL's checks (without replicating its
rather complex code).

The closest thing that fits the bill is Scalar::Util::reftype(),
which is what is used in this version, but only "required" if we detect
a ref in the candidate position in the call signature, so as to not
increase load times, at least when not using hash-storage.

Still, adding a dependency, even towards such a popular core module may
be undesirable sometimes... So I don't know.

Naturally, the regex check `( $_[0] =~  /HASH/ )` remains to be one of the
possible approaches, with its ow, set of pros and cons.

### CLOSING REMARKS

Again, sorry for the somewhat lengthy message :-)

---
Fixes: [#1][proper hash-storage support]
BasedOn: [#4][pull-request]
Unblocked by: [#3][implicit handlers/linkages]

References:

---

[#1]: https://github.com/perlancar/perl-Getopt-Long-More/issues/1
[#3]: https://github.com/perlancar/perl-Getopt-Long-More/issues/3
[#4]: https://github.com/perlancar/perl-Getopt-Long-More/pull/4
[#5]: https://github.com/perlancar/perl-Getopt-Long-More/issues/5

...